re-factor xgettext/extract command
