+++ packages/tools/src/commands/xgettext.ts
@@ -1,9 +1,18 @@
-import { copyFile, existsSync, readFileSync, writeFile } from 'fs';
-import { Command } from '../command';
-import { Textdomain, parseMoCatalog } from '@esgettext/runtime';
+import * as path from 'path';
+import { writeFileSync, readFileSync } from 'fs';
 import yargs from 'yargs';
-import * as mkdirp from 'mkdirp';
+
+import { Command } from '../command';
+import { Textdomain } from '@esgettext/runtime';
+import { Catalog, RenderOptions } from '../pot/catalog';
 import { Configuration } from '../configuration';
+import { CatalogProperties } from '../pot/catalog';
+import { PoParser } from '../parser/po';
+import { FilesCollector } from './xgettext/files-collector';
+import { Parser, ParserOptions } from '../parser/parser';
+import { TypeScriptParser } from '../parser/typescript';
+import { JavaScriptParser } from '../parser/javascript';
+import { Keyword } from '../pot/keyword';
 
 interface ExclusionCatalog {
 	[key: string]: Array<string>;
@@ -37,11 +46,15 @@ interface XGettextOptions {
 const gtx = Textdomain.getInstance('com.cantanea.esgettext-tools');
 
 export class XGettext implements Command {
-	private locales: Array<string>;
+	private catalog: Catalog;
+	private date: string | undefined;
+	private exclude: ExclusionCatalog;
 	private options: XGettextOptions;
 	private readonly configuration: Configuration;
 
-	constructor(configuration: Configuration) {
+	// The date is only passed for testing.
+	constructor(configuration: Configuration, date?: string) {
+		this.date = date;
 		this.configuration = configuration;
 	}
 
@@ -81,9 +94,7 @@ export class XGettext implements Command {
 				group: gtx._('Output file location:'),
 				alias: 'd',
 				type: 'string',
-				describe: gtx._(
-					'use NAME.po for output (instead of messages.po)',
-				),
+				describe: gtx._('use NAME.po for output (instead of messages.po)'),
 			},
 			output: {
 				group: gtx._('Output file location:'),
@@ -105,6 +116,8 @@ export class XGettext implements Command {
 				group: gtx._('Choice of input file language:'),
 				alias: 'L',
 				type: 'string',
+				choices: ['javascript', 'typescript'],
+				coerce: arg => arg.toLowerCase(),
 				describe: gtx._(
 					'recognise the specified language (JavaScript, TypeScript, HTML). Default is to auto-detect language based on filename extension.',
 				),
@@ -150,13 +163,13 @@ export class XGettext implements Command {
 				type: 'boolean',
 				describe: gtx._('extract all strings'),
 			},
-			'keyword': {
+			keyword: {
 				group: gtx._('Language specific options:'),
 				type: 'string',
 				describe: gtx._('look for WORD as an additional keyword'),
 				array: true,
 			},
-			'flag': {
+			flag: {
 				group: gtx._('Language specific options:'),
 				type: 'string',
 				describe: gtx._(
@@ -164,12 +177,10 @@ export class XGettext implements Command {
 				),
 				array: true,
 			},
-			'instance': {
+			instance: {
 				group: gtx._('Language specific options:'),
 				type: 'string',
-				describe: gtx._(
-					'only accept method calls of specified instance names',
-				),
+				describe: gtx._('only accept method calls of specified instance names'),
 				array: true,
 			},
 			'force-po': {
@@ -262,14 +273,360 @@ export class XGettext implements Command {
 		const options = argv as unknown as XGettextOptions;
 		this.options = options;
 		const conf = this.configuration;
+		const catalogProperties: CatalogProperties = { date: this.date };
 
+		catalogProperties.package = conf.package?.name;
+		catalogProperties.version = conf.package?.version;
+		catalogProperties.msgidBugsAddress = conf.package?.['msgid-bugs-address'];
+		catalogProperties.copyrightHolder = conf.package?.['copyright-holder'];
+
+		if (typeof options.output === 'undefined' && conf.package?.textdomain) {
+			if (typeof options.outputDir === 'undefined') {
+				options.output = conf.package.textdomain + '.pot';
+			} else {
+				options.output = path.join(
+					options.outputDir,
+					conf.package.textdomain + '.pot',
+				);
+			}
+		}
+
+		if (typeof options.packageName !== 'undefined') {
+			catalogProperties.package = options.packageName;
+		}
+		if (typeof options.packageVersion !== 'undefined') {
+			catalogProperties.version = options.version;
+		}
+		if (typeof options.msgidBugsAddress !== 'undefined') {
+			catalogProperties.msgidBugsAddress = options.msgidBugsAddress;
+		}
+
+		this.catalog = new Catalog(catalogProperties);
+
+		if (typeof options.language !== 'undefined') {
+			const language = options.language.toLowerCase();
+			if (!['javascript', 'typescript'].includes(language)) {
+				throw new Error(
+					gtx._x('language "{language}" unknown', {
+						language: this.options.language,
+					}),
+				);
+			}
+			this.options.language = language;
+		}
 	}
 
 	public run(argv: yargs.Arguments): Promise<number> {
 		this.init(argv);
 
 		return new Promise(resolve => {
-			resolve(0);
+			let exitCode = 0;
+
+			this.exclude = {} as ExclusionCatalog;
+			if (this.options.excludeFile) {
+				try {
+					if (!this.fillExclusionCatalog(this.options.excludeFile)) {
+						return 1;
+					}
+				} catch (e) {
+					const error = e as Error;
+					console.error(
+						gtx._x('{programName}: error: {message}', {
+							programName: this.options.$0,
+							message: error.message,
+						}),
+					);
+					return 1;
+				}
+			}
+
+			if (this.options.joinExisting) {
+				if (this.options.output === '-') {
+					console.error(
+						gtx._x(
+							'{programName}: error: --join-existing' +
+								' cannot be used, when output is written to stdout',
+							{
+								programName: this.options.$0,
+							},
+						),
+					);
+					return 1;
+				}
+
+				const parserOptions = this.getParserOptions();
+				const parser = new PoParser(this.catalog, parserOptions);
+				const filename: string = this.options.output;
+				try {
+					if (!parser.parse(this.readFile(filename), filename)) {
+						exitCode = 1;
+					}
+				} catch (msg) {
+					console.error(`${filename}: ${msg}`);
+					exitCode = 1;
+				}
+			}
+
+			let fileCollector;
+			try {
+				fileCollector = new FilesCollector(
+					this.options.filesFrom,
+					this.options._,
+				);
+			} catch (e) {
+				console.error(`${this.options.$0}: ${e}`);
+				return 1;
+			}
+
+			fileCollector.filenames.forEach(filename => {
+				try {
+					if (!this.parse(this.readFile(filename), filename)) {
+						exitCode = 1;
+					}
+				} catch (msg) {
+					if ('-' === filename) {
+						filename = gtx._('[standard input]');
+					}
+					console.error(`${filename}: ${msg}`);
+					exitCode = 1;
+				}
+			});
+
+			if (!exitCode) {
+				try {
+					this.output();
+				} catch (exception) {
+					console.error(
+						gtx._x('{programName}: {exception}', {
+							programName: this.options['$0'],
+							exception: exception as string,
+						}),
+					);
+					exitCode = 1;
+				}
+			}
+
+			return exitCode;
 		});
+	}
+
+	private parse(code: Buffer, filename: string): boolean {
+		let parser: Parser;
+		const parserOptions = this.getParserOptions();
+
+		if (typeof this.options.language !== 'undefined') {
+			parser = this.getParserByLanguage(this.options.language, parserOptions);
+		} else {
+			parser = this.getParserByFilename(filename, parserOptions);
+		}
+
+		return parser.parse(code, filename);
+	}
+
+	private readFile(filename: string): Buffer {
+		if ('-' === filename) {
+			return process.stdin.read() as Buffer;
+		}
+		const directories = this.options.directory || [''];
+
+		// Avoid ugly absolute paths.
+		const resolve = (dir: string, file: string): string => {
+			if (dir === '') {
+				return file;
+			} else {
+				return dir + path.sep + file;
+			}
+		};
+
+		for (let i = 0; i < directories.length - 1; ++i) {
+			try {
+				const fullName = resolve(directories[i], filename);
+				return readFileSync(fullName);
+			} catch (e) {
+				/* ignore */
+			}
+		}
+
+		return readFileSync(resolve(directories[directories.length - 1], filename));
+	}
+
+	private getParserByFilename(
+		filename: string,
+		parserOptions: ParserOptions,
+	): Parser {
+		let parser: Parser;
+		const ext = path.extname(filename);
+
+		switch (ext.toLocaleLowerCase()) {
+			case '.ts':
+			case '.tsx':
+				parser = new TypeScriptParser(this.catalog, parserOptions);
+				break;
+			case '.js':
+			case '.jsx':
+				parser = new JavaScriptParser(this.catalog, parserOptions);
+				break;
+			case '.po':
+			case '.pot':
+				parser = new PoParser(this.catalog, parserOptions);
+				break;
+			default:
+				if ('-' === filename) {
+					this.warn(
+						gtx._(
+							'language for standard input is unknown without' +
+								' option "--language"; will try JavaScript',
+						),
+					);
+				} else {
+					this.warn(
+						gtx._x(
+							'file "{filename}" extension "{extname}"' +
+								' is unknown; will try JavaScript instead',
+							{
+								filename,
+								extname: ext,
+							},
+						),
+					);
+				}
+				parser = new JavaScriptParser(this.catalog, parserOptions);
+				break;
+		}
+
+		return parser;
+	}
+
+	private getParserByLanguage(
+		language: string,
+		parserOptions: ParserOptions,
+	): Parser {
+		let parser: Parser;
+		switch (language.toLocaleLowerCase()) {
+			case 'typescript':
+				parser = new TypeScriptParser(this.catalog, parserOptions);
+				break;
+			case 'javascript':
+			default:
+				parser = new JavaScriptParser(this.catalog, parserOptions);
+				break;
+		}
+
+		return parser;
+	}
+
+	private warn(msg: string): void {
+		console.warn(
+			gtx._x('{programName}: warning: {msg}', {
+				msg,
+				programName: this.options.$0,
+			}),
+		);
+	}
+
+	private output(): void {
+		Object.keys(this.exclude)
+			.filter(msgctxt =>
+				Object.prototype.hasOwnProperty.call(this.exclude, msgctxt),
+			)
+			.forEach(msgctxt => {
+				this.exclude[msgctxt].forEach(msgid => {
+					this.catalog.deleteEntry(msgid, msgctxt);
+				});
+			});
+
+		if (!this.options.forcePo) {
+			if (this.catalog.entries.length < 2) {
+				return;
+			}
+		}
+
+		const renderOptions: RenderOptions = {};
+
+		if (typeof this.options.width !== 'undefined') {
+			renderOptions.width = this.options.width;
+		}
+
+		const po = this.catalog.toString(renderOptions);
+
+		if (this.options.output === '-') {
+			process.stdout.write(po);
+			return;
+		}
+
+		let filename;
+		if (typeof this.options.output === 'undefined') {
+			const domain =
+				typeof this.options.defaultDomain === 'undefined'
+					? 'messages'
+					: this.options.defaultDomain;
+			filename = `${domain}.po`;
+		} else {
+			filename = this.options.output;
+		}
+
+		const outputDir =
+			typeof this.options.outputDir === 'undefined'
+				? ''
+				: this.options.outputDir;
+
+		writeFileSync(path.join(outputDir, filename), po);
+	}
+
+	private fillExclusionCatalog(catalogs: Array<string>): boolean {
+		const catalog = new Catalog();
+		const parser = new PoParser(catalog);
+		let success = true;
+
+		catalogs.forEach(filename => {
+			if (!parser.parse(this.readFile(filename), filename)) {
+				success = false;
+			}
+		});
+
+		if (!success) {
+			return false;
+		}
+
+		catalog.deleteEntry('');
+
+		catalog.entries.forEach(entry => {
+			if (
+				typeof this.exclude[entry.properties.msgctxt as string] === 'undefined'
+			) {
+				this.exclude[entry.properties.msgctxt as string] = [];
+			}
+			if (entry.properties.msgid) {
+				this.exclude[entry.properties.msgctxt as string].push(
+					entry.properties.msgid,
+				);
+			}
+		});
+
+		return true;
+	}
+
+	private getParserOptions(): ParserOptions {
+		const parserOptions: ParserOptions = (({
+			fromCode,
+			addComments,
+			addAllComments,
+			extractAll,
+		}): ParserOptions => ({
+			fromCode,
+			addComments,
+			addAllComments,
+			extractAll,
+		}))(this.options);
+
+		if (this.options.keyword) {
+			const cookedKeywords = new Array<Keyword>();
+			this.options.keyword.forEach(raw => {
+				cookedKeywords.push(Keyword.from(raw));
+			});
+			parserOptions.keyword = cookedKeywords;
+		}
+
+		return parserOptions;
 	}
 }
