+++ packages/tools/src/commands/xgettext/file-resolver.spec.ts
