+++ packages/tools/src/commands/xgettext/file-resolver.ts
