+++ packages/tools/src/commands/xgettext/files-collector.spec.ts
