+++ packages/tools/src/commands/xgettext/files-collector.ts
