+++ packages/tools/src/commands/xgettext/xgettext.spec.ts
